refactor: 그룹 스웨거 파일 분할
